feat: renomear módulo rom-discovery para kid-discovery, adicionar testes para a classe Rom e implementar nova função para obter bytes de células em uma imagem Indexed4
