+++ packages/kid-util/src/index.ts
@@ -15,5 +15,5 @@
 export * from './kid-rom'
 export * from './kid-utils'
 export * from './kid-resources'
-export * from './rom-discovery'
+export * from './kid-discovery'
 export * from './kid-graphics'
